# backend/openapi/openapi.yml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Toe Beans
  description: API reference of Toe Beans
servers:
  - url: http://localhost:80
    description: local
  - url: will be written
    description: production
paths:
...
  /user:
    post:
      description: register user info
      operationId: registerUser
      tags:
        - user
      requestBody:
        $ref: '#/components/requestBodies/registerUser'
      responses:
        "200":
          $ref: '#/components/responses/simpleSuccess'
        "400":
          $ref: '#/components/responses/badRequest'
        "405":
          $ref: '#/components/responses/notAllowedMethod'
        "500":
          $ref: '#/components/responses/internalServerError'
    get:
      description: get a user info
      operationId: getUser
      tags:
        - user
      security:
        - bearerAuth: []
      parameters:
        - name: user_name
          in: query
          description: user name
          required: true
          schema:
            type: string
            example: user1
      responses:
        "200":
          $ref: '#/components/responses/getUser'
        "400":
          $ref: '#/components/responses/badRequest'
        "401":
          $ref: '#/components/responses/unauthorized'
        "404":
          $ref: '#/components/responses/notFound'
        "405":
          $ref: '#/components/responses/notAllowedMethod'
        "500":
          $ref: '#/components/responses/internalServerError'
    put:
      description: update user info
      operationId: updateUser
      tags:
        - user
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/updateUser'
      responses:
        "200":
          $ref: '#/components/responses/simpleSuccess'
        "400":
          $ref: '#/components/responses/badRequest'
        "401":
          $ref: '#/components/responses/unauthorized'
        "403":
          $ref: '#/components/responses/forbidden'
        "405":
          $ref: '#/components/responses/notAllowedMethod'
        "500":
          $ref: '#/components/responses/internalServerError'
    delete:
      description: delete user info
      operationId: deleteUser
      tags:
        - user
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: '#/components/responses/simpleSuccess'
        "401":
          $ref: '#/components/responses/unauthorized'
        "403":
          $ref: '#/components/responses/forbidden'
        "405":
          $ref: '#/components/responses/notAllowedMethod'
        "500":
          $ref: '#/components/responses/internalServerError'
...
  /comments:
    get:
      description: get comments
      operationId: getComments
      tags:
        - comment
      security:
        - bearerAuth: []
      parameters:
        - name: posting_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          $ref: '#/components/responses/getComments'
        "400":
          $ref: '#/components/responses/badRequest'
        "401":
          $ref: '#/components/responses/unauthorized'
        "405":
          $ref: '#/components/responses/notAllowedMethod'
        "500":
          $ref: '#/components/responses/internalServerError'
...
  /notifications:
    get:
      description: get notifications
      operationId: getNotifications
      tags:
        - notification
      security:
        - bearerAuth: []
      parameters:
        - name: user_name
          description: user name
          in: query
          required: true
          schema:
            type: string
          explode: true
      responses:
        "200":
          $ref: '#/components/responses/getNotifications'
        "400":
          $ref: '#/components/responses/badRequest'
        "401":
          $ref: '#/components/responses/unauthorized'
        "403":
          $ref: '#/components/responses/forbidden'
        "405":
          $ref: '#/components/responses/notAllowedMethod'
        "500":
          $ref: '#/components/responses/internalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
...
    registerUser:
      description: register user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/requestRegisterUser'
    updateUser:
      description: update user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/requestUpdateUser'
...
  responses:
    getUser:
      description: get user
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseGetUser'
...
    getComments:
      description: get comments
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseGetComments'
    getNotifications:
      description: get notifications
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseGetNotifications'
    simpleSuccess:
      description: '200'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseSimpleSuccess'
    badRequest:
      description: '400'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseBadRequest'
    unauthorized:
      description: '401'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseUnauthorized'
    forbidden:
      description: '403'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseForbidden'
    notFound:
      description: '404'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseNotFound'
    notAllowedMethod:
      description: '405'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseNotAllowedMethod'
    internalServerError:
      description: '500'
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/responseInternalServerError'
  schemas:
...
    requestRegisterUser:
      type: object
      properties:
        email:
          type: string
          description: email
          example: user1@gmail.com
        user_name:
          type: string
          description: user_name
          example: user1
        password:
          type: string
          description: password
          example: Password1234
      required:
        - email
        - user_name
        - password
    requestUpdateUser:
      type: object
      properties:
        password:
          type: string
          description: password
          example: Password1234
        icon:
          description: icon image file of base64 encoded and zipped
          type: string
          format: byte
          example: 'GEsDBBQACAAIAJhjzE4AAAAAAAAAAAAAAAASABAAaU9TIOOBrueUu+WDjzIucG5nVVgMAKTALl1wcQBd9gEUAIy8B'
        self_introduction:
          type: string
          description: self introduction
          example: 'Hello'
...
    responseGetUser:
      description: get user
      type: object
      properties:
        user_name:
          description: user name
          type: string
          example: user1
        icon:
          description: icon
          type: string
          example: icon url
        posting_count:
          description: the total count of posting
          type: integer
          format: int64
          example: 1
        like_count:
          description: the total count of like
          type: integer
          format: int64
          example: 10
        liked_count:
          description: the total count of liked
          type: integer
          format: int64
          example: 20
        created_at:
          description: the datetime when the account is created
          type: string
          format: date-time
          example: '2020-01-01T00:00:00Z'
      required:
        - user_name
        - icon
        - posting_count
        - like_count
        - liked_count
        - created_at
...
    responseGetComments:
      description: get comments
      type: object
      properties:
        posting_id:
          description: posting id
          type: integer
          format: int64
          example: 1
        comments:
          description: list of comment
          type: array
          items:
            $ref: '#/components/schemas/responseGetComment'
    responseGetComment:
      type: object
      properties:
        comment_id:
          description: comment id
          type: integer
          format: int64
          example: 1
        user_name:
          type: string
          description: user_name
          example: user1
        commented_at:
          description: commented datetime with TZ. This means created_at in postings table.
          type: string
          format: date-time
          example: '2020-01-01T00:00:00Z'
        comment:
          type: string
          description: the content of comment
          example: This is a sample comment.
      required:
        - comment_id
        - user_name
        - commented_at
        - comment
    responseGetNotifications:
      description: get notifications
      type: object
      properties:
        visited_name:
          description: acted user name
          type: string
          example: user1
        actions:
          description: actions
          type: array
          items:
            $ref: '#/components/schemas/responseGetNotification'
    responseGetNotification:
      type: object
      properties:
        visitor_name:
          description: acting user name
          type: string
          example: user1
        posting_id:
          description: only when action is like or comment
          type: integer
          format: int64
          example: 1
        comment_id:
          description: only when action is comment
          type: integer
          format: int64
          example: 1
        action_type:
          description: action type
          type: string
          enum:
            - 'like'
            - 'comment'
            - 'follow'
        created_at:
          description: datetime with TZ
          type: string
          format: date-time
          example: '2020-01-01T00:00:00Z'
      required:
        - visited_name
        - visitor_name
        - action
        - created_at
    responseSimpleSuccess:
      description: Success
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 200
        message:
          description: message
          type: string
          example: 'success'
      required:
        - status
        - message
    responseBadRequest:
      description: Bad Reqeust
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 400
        message:
          description: message
          type: string
          example: 'Bad Request'
      required:
        - status
        - message
    responseUnauthorized:
      description: Unauthorized
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 401
        message:
          description: message
          type: string
          example: 'Unauthorized'
      required:
        - status
        - message
    responseForbidden:
      description: Forbidden
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 403
        message:
          description: message
          type: string
          example: 'Forbidden'
      required:
        - status
        - message
    responseNotFound:
      description: Not Found
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 404
        message:
          description: message
          type: string
          example: 'Not Found'
      required:
        - status
        - message
    responseNotAllowedMethod:
      description: Not Allowed Method
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 405
        message:
          description: message
          type: string
          example: 'Not Allowed Method'
      required:
        - status
        - message
    responseInternalServerError:
      description: Internal Server Error
      type: object
      properties:
        status:
          description: status
          type: integer
          format: int16
          example: 500
        message:
          description: message
          type: string
          example: 'Internal Server Error'
      required:
        - status
        - message
...